Supporting multiple routes. Dynamically assigned to cars with FollowRoute action.
Former-commit-id: 40e843c75c596a2318242eb1d28243f8c1ecb42e

+++ EnvironmentSimulator/ScenarioEngineMini/Conditions/Condition.cpp
@@ -17,7 +17,7 @@ void Condition::identifyConditionType(OSCCondition &condition)
 		this->conditionType = "TimeHeadway";
 
 		// Triggering entities		(TriggeringEntities -> Entity -> name="Ego")
-		this->N = condition.ByEntity.TriggeringEntities.Entity.size();
+		this->N = (int)condition.ByEntity.TriggeringEntities.Entity.size();
 		this->triggeringEntityPos.resize(N);
 		this->triggeringEntities.resize(N);
